+++ packages/solidjs-integration/src/type-generator.ts
@@ -12,9 +12,17 @@ import { Options } from "./types";
 export function generateSolidJsTypes(manifest: any, options: Options) {
   options = getOptions(options);
 
-  const components = getComponents(manifest, options.exclude).filter(x => x.tagName);
+  const components = getComponents(manifest, options.exclude).filter(
+    (x) => x.tagName
+  );
   const template = getTypeTemplate(components, options);
-  const outputPath = saveFile(options.outdir!, options.fileName!, template, "typescript", 120);
+  const outputPath = saveFile(
+    options.outdir!,
+    options.fileName!,
+    template,
+    "typescript",
+    120
+  );
   logBlue(`[solidjs-type-generator] - Generated "${outputPath}".`);
 }
 
@@ -31,7 +39,9 @@ function getOptions(options: Options) {
 function getEventTypes(component: Component, componentNames: string[]) {
   const types = component.events
     ?.map((e) => {
-      const eventType = e.type?.text.replace("[]", "").replace(" | undefined", "");
+      const eventType = e.type?.text
+        .replace("[]", "")
+        .replace(" | undefined", "");
       return eventType &&
         !EXCLUDED_TYPES.includes(eventType) &&
         !componentNames.includes(eventType) &&